Update multisig signature tx parser function  (#25)
- Support inner transactions after the origin data
- Update signing byte length for this tx type
- Add test for multisig signature transfer and provision namespace tx in test_nem.py
- Update mosaic num screen in transfer tx

+++ src/nem/parse/nem_parse.h
@@ -28,7 +28,7 @@ typedef struct result_t {
 
 typedef struct parse_context_t {
     uint8_t version;
-    uint16_t transactionType;
+    uint32_t transactionType;
     uint8_t *data;
     result_t result;
     uint32_t length;
